Moved joinmarket protocol logic for privmsg send to message channel class. IRC level processing (chunking, headers, delimiters) still in IRCMessageChannel. Encryption and command string logic moved.
